Split Replication interface into encoder/parser
